requirements.txt
python-dotenv==0.21.1
toolz==0.12.0
jupyter==1.0.0